apps/api-python/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==7.4.2
httpx==0.24.1
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6